第一个 postgresql 版本，注意目前还不支持定时任务

+++ ruoyi-fastapi-backend/module_admin/entity/do/config_do.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from sqlalchemy import Column, DateTime, Integer, String
+from sqlalchemy import Column, DateTime, Integer, String, BigInteger
 from config.database import Base
 
 
@@ -10,7 +10,7 @@ class SysConfig(Base):
 
     __tablename__ = 'sys_config'
 
-    config_id = Column(Integer, primary_key=True, autoincrement=True, comment='参数主键')
+    config_id = Column(BigInteger, primary_key=True, autoincrement=True, comment='参数主键')
     config_name = Column(String(100), nullable=True, default='', comment='参数名称')
     config_key = Column(String(100), nullable=True, default='', comment='参数键名')
     config_value = Column(String(500), nullable=True, default='', comment='参数键值')
