+++ ruoyi-fastapi-backend/module_admin/entity/do/notice_do.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from sqlalchemy import Column, DateTime, Integer, LargeBinary, String
+from sqlalchemy import Column, DateTime, Integer, LargeBinary, String, BigInteger
 from config.database import Base
 
 
@@ -10,7 +10,7 @@ class SysNotice(Base):
 
     __tablename__ = 'sys_notice'
 
-    notice_id = Column(Integer, primary_key=True, autoincrement=True, comment='公告ID')
+    notice_id = Column(BigInteger, primary_key=True, autoincrement=True, comment='公告ID')
     notice_title = Column(String(50, collation='utf8_general_ci'), nullable=False, comment='公告标题')
     notice_type = Column(String(1, collation='utf8_general_ci'), nullable=False, comment='公告类型（1通知 2公告）')
     notice_content = Column(LargeBinary, comment='公告内容')
