+++ ruoyi-fastapi-backend/module_admin/entity/do/role_do.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from sqlalchemy import Column, DateTime, Integer, String
+from sqlalchemy import Column, DateTime, Integer, String, BigInteger
 from config.database import Base
 
 
@@ -10,7 +10,7 @@ class SysRole(Base):
 
     __tablename__ = 'sys_role'
 
-    role_id = Column(Integer, primary_key=True, autoincrement=True, comment='角色ID')
+    role_id = Column(BigInteger, primary_key=True, autoincrement=True, comment='角色ID')
     role_name = Column(String(30, collation='utf8_general_ci'), nullable=False, comment='角色名称')
     role_key = Column(String(100, collation='utf8_general_ci'), nullable=False, comment='角色权限字符串')
     role_sort = Column(Integer, nullable=False, comment='显示顺序')
@@ -37,8 +37,8 @@ class SysRoleDept(Base):
 
     __tablename__ = 'sys_role_dept'
 
-    role_id = Column(Integer, primary_key=True, nullable=False, comment='角色ID')
-    dept_id = Column(Integer, primary_key=True, nullable=False, comment='部门ID')
+    role_id = Column(BigInteger, primary_key=True, nullable=False, comment='角色ID')
+    dept_id = Column(BigInteger, primary_key=True, nullable=False, comment='部门ID')
 
 
 class SysRoleMenu(Base):
@@ -48,5 +48,5 @@ class SysRoleMenu(Base):
 
     __tablename__ = 'sys_role_menu'
 
-    role_id = Column(Integer, primary_key=True, nullable=False, comment='角色ID')
-    menu_id = Column(Integer, primary_key=True, nullable=False, comment='菜单ID')
+    role_id = Column(BigInteger, primary_key=True, nullable=False, comment='角色ID')
+    menu_id = Column(BigInteger, primary_key=True, nullable=False, comment='菜单ID')
